mirrorGoal/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
channels==4.2.2
channels_redis==4.2.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-shortuuidfield==0.1.3
django-taggit==6.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.5.3
kombu==5.5.3
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.0.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
daphne==4.1.2
gunicorn==21.2.0
